rename spectrum_time
