+++ source/consumers/TimeSpectrum.h
